requirements.txt
aiohttp~=3.6.2
asyncio
prompt-toolkit<3.0.0
pygments
PyPubSub
python-dotenv